+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_correct_test_file_name/examples/example.dart
@@ -1,0 +1,4 @@
+// LINT
+void main() {
+  print('Hello');
+}