+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_unnecessary_type_assertions/examples/example_with_not_is.dart
@@ -26,3 +26,8 @@ void checkSameOrInheritor() {
   final Object b;
   final check = b is! int;
 }
+
+void check<T>() {
+  final T b;
+  final check = b is! int;
+}